Nuevos cambios realizados a peticion de mejora continua

+++ app/src/main/java/com/example/sm_tubo_plast/genesys/BEAN/ReportePedidoDetallePDF.kt
@@ -17,7 +17,7 @@ class ReportePedidoDetallePDF(
     companion object{
         //crear un metodo para que ordene los datos de la lista, por oc_numero y item, donde reciba una lista
         fun orderOcNumeroAndItem(lista: ArrayList<ReportePedidoDetallePDF>): List<ReportePedidoDetallePDF>{
-            var listaOrdenada = lista.sortedWith(compareBy({it.oc_numero}, {it.item}))
+            var listaOrdenada = lista.sortedWith(compareBy({it.oc_numero}, {it.item.toInt()}))
             return listaOrdenada
         }
     }
